+++ R/makeSingleExpertPlot.R
@@ -1,7 +1,7 @@
 makeSingleExpertPlot <-
 function(fit, d = "best", pl = -Inf, pu = Inf,
          ql = NA, qu = NA, sf = 3, ex = 1,
-         lwd = 1, xlab, ylab){
+         lwd = 1, xlab, ylab, percentages){
   
 	if(d == "best"){
 		ssq <- fit$ssq[ex, is.na(fit$ssq[ex,])==F]
@@ -223,7 +223,6 @@ function(fit, d = "best", pl = -Inf, pu = Inf,
 	df1 <- data.frame(x = x, fx = fx)
 	p1 <- ggplot(df1, aes(x = x, y = fx)) +
 	  geom_line(size = lwd) +
-	  xlim(pl, pu) + 
 	  labs(title = dist.title, x = xlab, y = ylab )+
 	  theme(plot.title = element_text(hjust = 0.5))
 	if(is.na(ql) == F  ){
@@ -238,5 +237,13 @@ function(fit, d = "best", pl = -Inf, pu = Inf,
 	                         fill = "red",
 	                         alpha = 0.5)
 	}
+	
+	if(percentages){
+	  p1 <- p1 + scale_x_continuous(labels = scales::percent,  
+	                                limits = c(pl, pu))
+	}else{
+	  p1 <- p1 + xlim(pl, pu)
+	}
+	
 	p1
 }