+++ R/plotTertiles.R
@@ -9,6 +9,8 @@
 #' row is the 33rd percentile, 2nd row is the median, last row is the 66th percentile.
 #' @param lower a vector of lower plausible limits: one per expert
 #' @param upper a vector of upper plausible limits: one per expert
+#' @param fs font size to be used in the plot
+#' @param percentages set to \code{TRUE} to use percentages on the x-axis
 #'
 #
 
@@ -26,7 +28,7 @@
 #' }
 #' @export
 
-plotTertiles <- function(vals, lower, upper, fs = 12){
+plotTertiles <- function(vals, lower, upper, fs = 12, percentages = FALSE){
   
   low <- L <- T1 <- M <- T2 <- U <- enumber <- NULL # hack to pass CRAN check
   
@@ -40,14 +42,18 @@ plotTertiles <- function(vals, lower, upper, fs = 12){
   colnames(df1) <- c("L", "T1", "M", "T2", "U")
   df1$expert <- expert
   df1$enumber <- n.experts:1
-  ggplot(df1, aes(x = low, y = expert)) +
+  p1 <- ggplot(df1, aes(x = low, y = expert)) +
     geom_segment(aes(yend = expert, x=L, xend = T1), lwd = 10, col = cols[1])+
     geom_segment(aes(yend = expert, x=T1, xend = T2), lwd = 10, col = cols[2])+
     geom_segment(aes(yend = expert, x=T2, xend = U), lwd = 10, col = cols[3])+
     geom_segment(aes(y = enumber -0.15, yend = enumber + 0.15, x = M, xend =M),
                  lwd = 1, linetype = "dashed")+
     labs(x = "X") +
-    theme(text = element_text(size = fs))
+    theme(text = element_text(size = fs)) 
+  if(percentages){
+    p1 <- p1 + scale_x_continuous(labels = scales::percent)
+  }
+  p1  
 }